# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: OmniCLI Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /files/find:
    get:
      operationId: findFiles
      tags: [files]
      summary: Find files matching a pattern
      parameters:
        - name: pattern
          in: query
          schema:
            type: string
        - name: path
          in: query
          schema:
            type: string
            default: "."
        - name: type
          in: query
          schema:
            type: string
            enum: [any, f, d, l]
            default: any
        - name: size
          in: query
          schema:
            type: string
        - name: modified
          in: query
          schema:
            type: string
        - name: maxDepth
          in: query
          schema:
            type: integer
        - name: regex
          in: query
          schema:
            type: boolean
            default: false
        - name: limit
          in: query
          schema:
            type: integer
            default: 200
      responses:
        "200":
          description: Found files
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FindResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /files/hash:
    get:
      operationId: hashFile
      tags: [files]
      summary: Compute the hash of a file
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
        - name: algo
          in: query
          schema:
            type: string
            enum: [blake3, sha256, md5]
            default: blake3
      responses:
        "200":
          description: Hash result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HashResult"
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /files/stats:
    get:
      operationId: getFileStats
      tags: [files]
      summary: Get file system stats for a directory
      parameters:
        - name: path
          in: query
          schema:
            type: string
            default: "."
      responses:
        "200":
          description: Directory statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileStats"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Archive ─────────────────────────────────────────────────────────────────
...
  /dev/hash:
    post:
      operationId: devHash
      tags: [dev]
      summary: Hash text or file content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DevHashInput"
      responses:
        "200":
          description: Hash result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DevHashResult"
  /dev/json:
    post:
      operationId: devJson
      tags: [dev]
      summary: Format, minify, or query JSON
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DevJsonInput"
      responses:
        "200":
          description: JSON result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DevJsonResult"
        "400":
          description: Invalid JSON
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /dev/base64:
    post:
      operationId: devBase64
      tags: [dev]
      summary: Encode or decode base64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DevBase64Input"
      responses:
        "200":
          description: Base64 result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DevBase64Result"
  /dev/uuid:
    get:
      operationId: devUuid
      tags: [dev]
      summary: Generate a UUID v4
      parameters:
        - name: count
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Generated UUIDs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DevUuidResult"
  /dev/regex:
    post:
      operationId: devRegex
      tags: [dev]
      summary: Test a regex pattern against text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DevRegexInput"
      responses:
        "200":
          description: Regex match results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DevRegexResult"
        "400":
          description: Invalid regex
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /dev/jwt:
    post:
      operationId: devJwt
      tags: [dev]
      summary: Decode and inspect a JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DevJwtInput"
      responses:
        "200":
          description: Decoded JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DevJwtResult"
        "400":
          description: Invalid JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Dashboard ──────────────────────────────────────────────────────────────
...
    FileEntry:
      type: object
      required: [path, fileType, sizeBytes, modified]
      properties:
        path:
          type: string
        fileType:
          type: string
        sizeBytes:
          type: integer
        modified:
          type: integer
    FindResponse:
      type: object
      required: [entries, total, durationMs]
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/FileEntry"
        total:
          type: integer
        durationMs:
          type: number
    HashResult:
      type: object
      required: [path, algo, digest, sizeBytes, durationMs]
      properties:
        path:
          type: string
        algo:
          type: string
        digest:
          type: string
        sizeBytes:
          type: integer
        durationMs:
          type: number
    FileStats:
      type: object
      required: [path, totalFiles, totalDirs, totalSizeBytes, largestFile, typeBreakdown]
      properties:
        path:
          type: string
        totalFiles:
          type: integer
        totalDirs:
          type: integer
        totalSizeBytes:
          type: integer
        largestFile:
          type: ["string", "null"]
        largestFileSizeBytes:
          type: ["integer", "null"]
        typeBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/FileTypeCount"
    FileTypeCount:
      type: object
      required: [extension, count, totalSizeBytes]
      properties:
        extension:
          type: string
        count:
          type: integer
        totalSizeBytes:
          type: integer

    # ── Archive ─────────────────────────────────────────────────────────────────
...
    DevHashInput:
      type: object
      required: [input]
      properties:
        input:
          type: string
          minLength: 1
        algo:
          type: string
          enum: [blake3, sha256, md5, sha1]
          default: sha256
    DevHashResult:
      type: object
      required: [input, algo, digest]
      properties:
        input:
          type: string
        algo:
          type: string
        digest:
          type: string
    DevJsonInput:
      type: object
      required: [input]
      properties:
        input:
          type: string
          minLength: 1
        action:
          type: string
          enum: [pretty, minify, validate]
          default: pretty
        query:
          type: ["string", "null"]
    DevJsonResult:
      type: object
      required: [valid, output]
      properties:
        valid:
          type: boolean
        output:
          type: string
        error:
          type: ["string", "null"]
    DevBase64Input:
      type: object
      required: [input]
      properties:
        input:
          type: string
          minLength: 1
        decode:
          type: boolean
          default: false
    DevBase64Result:
      type: object
      required: [input, output, decoded]
      properties:
        input:
          type: string
        output:
          type: string
        decoded:
          type: boolean
    DevUuidResult:
      type: object
      required: [uuids]
      properties:
        uuids:
          type: array
          items:
            type: string
    DevRegexInput:
      type: object
      required: [pattern, text]
      properties:
        pattern:
          type: string
          minLength: 1
        text:
          type: string
        flags:
          type: string
    DevRegexResult:
      type: object
      required: [pattern, matched, matchCount, matches]
      properties:
        pattern:
          type: string
        matched:
          type: boolean
        matchCount:
          type: integer
        matches:
          type: array
          items:
            $ref: "#/components/schemas/RegexMatch"
        error:
          type: ["string", "null"]
    RegexMatch:
      type: object
      required: [match, start, end]
      properties:
        match:
          type: string
        start:
          type: integer
        end:
          type: integer
        groups:
          type: array
          items:
            type: string
    DevJwtInput:
      type: object
      required: [token]
      properties:
        token:
          type: string
          minLength: 1
    DevJwtResult:
      type: object
      required: [valid, header, payload]
      properties:
        valid:
          type: boolean
        header:
          type: string
        payload:
          type: string
        signature:
          type: string
        isExpired:
          type: ["boolean", "null"]
        expiresAt:
          type: ["string", "null"]
        error:
          type: ["string", "null"]