# lib/api-zod/src/generated/types/devJsonInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * OmniCLI Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { DevJsonInputAction } from './devJsonInputAction';

export interface DevJsonInput {
  /** @minLength 1 */
  input: string;
  action?: DevJsonInputAction;
  /** @nullable */
  query?: string | null;
}