# lib/api-zod/src/generated/types/devJwtInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * OmniCLI Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface DevJwtInput {
  /** @minLength 1 */
  token: string;
}