# lib/api-zod/src/generated/types/fileStats.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * OmniCLI Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { FileTypeCount } from './fileTypeCount';

export interface FileStats {
  path: string;
  totalFiles: number;
  totalDirs: number;
  totalSizeBytes: number;
  /** @nullable */
  largestFile: string | null;
  /** @nullable */
  largestFileSizeBytes?: number | null;
  typeBreakdown: FileTypeCount[];
}